==> 3_dyn_static/analysis/4.5_calculating_zscores_test.Rmd <==
---
title: "4.5_calculating_zscores_test"
author: "Jenny Richmond"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

https://rpubs.com/jenrichmond/emgz 


### load packages 
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(skimr)

options(scipen = 999) #set no scientific notation
```

### The process
- first, make data wide using pivot_wider so that brow and cheek are in columns
- use scale() to transform RMS scores into z scores for each participant.
- then make bin wide so that each row represents a single trial and bins 1-6 are in different columns
- make new column for each bin, calculating the difference between that binZ and baselineZ.
- make z diff scores long again

# how to: calculate zscores

Normally, to center a variable, you would subtract the mean of all data points from each individual data point. With scale(), this can be accomplished in one simple call.

Center variable A using the scale() function

`scale(A, center = TRUE, scale = FALSE)`

You can generate z-scores for variable A using the scale() function

`scale(A, center = TRUE, scale = TRUE)`



## read in data 

```{r message=FALSE, warning=FALSE}
data <- read_csv(here("data", "combined", "2_clean_no_bl_outliers.csv"))
                      
```

# Testing z score process

In Sophie's analysis she separated her data into baseline and bins and calculated z scores separately before joining them back together. For Sahar and Riley's analysis, we have calcualted z scores with baseline and bins together. 

This rmd tests whether separate or togehter makes any difference to the distribution of z scores. 

> NOTE all dataframes created from the together process have a "t" in front, all df created via the separate process have a "s" in front, to allow for comparisons

# **TOGETHER**

### 1 pp first

```{r}
tpp403 <- data %>%
  filter(pp_no == "pp403")

# make it wide 
twidepp403 <- tpp403 %>%
  pivot_wider(names_from = muscle, values_from = rms)

# create new variables called zbrow and zcheek using the scale function and specifying center = TRUE and scale = TRUE

twidepp403$zbrow = scale(twidepp403$brow, center = TRUE, scale = TRUE)
twidepp403$zcheek = scale(twidepp403$cheek, center = TRUE, scale = TRUE)

# Check that the mean of the new zbrow and zcheek columns is 0 - this is what we want... 
summary(twidepp403)

```


##  all the data

make data wide, use mutate to make new z variables
```{r}

tdata_zscore <- data %>%
  pivot_wider(names_from = muscle, 
              values_from = rms) %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))
  
summary(tdata_zscore)
```

> note: can also make new z-variables manually: 

> `data_z_manual <- data_wide %>% 
group_by(pp_no) %>%
mutate(Zbrow = (brow - mean(brow))/sd(brow)) %>%
mutate(Zcheek = (cheek - mean(cheek))/sd(cheek))`



##  Difference Scores 

Need to calculate difference scores from baseline. Need to make BIN wide to allow for calculations across columns. Its a bit difficult to do that for both muscles at the same time, so lets separate and work out the difference scores for brow and cheek separately.

### BROW FIRST

#### Add a muscle column 

Just to make things easier to join back together later
```{r}

tbrow_z <- tdata_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, condition, emotion, bin, trial, muscle, Zbrow) 
```

#### Make the bin column wide 

... and rename bin_0 as BL (i.e. baseline)
```{r}
tbrow_z_wide <- tbrow_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zbrow) %>%
  rename(BL = bin_0)

```

#### mutate() diff scores 

Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

```{r}
tbrow_z_diff <- tbrow_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL) %>%
  select(-BL, - starts_with("bin"))
```

#### Make data long again 

```{r}
tbrow_z_diff_long <- tbrow_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin6)
  
```

#### checking on missingness
```{r}
library(naniar)

vis_miss(tbrow_z_diff_long)
```



### NOW CHEEK

#### add a muscle column 

...to make things easier to join back together later
```{r}

tcheek_z <- tdata_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, condition, emotion, bin, trial, muscle, Zcheek) 

```

#### Make the bin column wide 

...and rename bin_0 as BL (i.e. baseline)
```{r}

tcheek_z_wide <- tcheek_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zcheek) %>%
  rename(BL = bin_0)
```

#### mutate() diff scores 

- Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

```{r}
tcheek_z_diff <- tcheek_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL) %>%
  select(-BL, - starts_with("bin"))
```
#### make data long again 

```{r}

tcheek_z_diff_long <- tcheek_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin6)
  
```


# BIND CHEEK AND BROW TOGETHER

```{r}
tzdiff_cheek_brow <- bind_rows(tcheek_z_diff_long, tbrow_z_diff_long) %>%
  arrange(pp_no, emotion, trial, muscle, bin)
```

# PLOT distribution of zdiff

```{r}

tzdiff_cheek_brow %>%
  group_by(muscle, emotion) %>%
  summarise(meanZdiff = mean(Zdiff, na.rm= TRUE), 
            sdZdiff = sd(Zdiff, na.rm=TRUE))


tog_plot <- tzdiff_cheek_brow %>%
  ggplot(aes(x = bin, y = Zdiff)) +
  geom_jitter() +
  facet_grid(vars(muscle), vars(emotion)) + 
  labs(title = "Zdiff bl and bin together")
```

# write to csv 
```{r}
tzdiff_cheek_brow %>%
  write_csv(here("data", "combined", "t4_zdiff_clean.csv"))

beepr::beep(1)
```


# **SEPARATE**

## read in data 

```{r message=FALSE, warning=FALSE}
data <- read_csv(here("data", "combined", "2_clean_no_bl_outliers.csv"))
                      
```

### 1 pp first

```{r}
spp403 <- data %>%
  filter(pp_no == "pp403")

# make it wide 
swidepp403 <- spp403 %>%
  pivot_wider(names_from = muscle, values_from = rms)

# filter for bl and bin 

swidepp403_bl <- swidepp403 %>%
  filter(bin == "bin_0")

swidepp403_bin <- swidepp403 %>%
  filter(bin != "bin_0")


# for bl and bins separately, create new variables called zbrow and zcheek using the scale function and specifying center = TRUE and scale = TRUE

# bl first
swidepp403_bl$zbrow = scale(swidepp403_bl$brow, center = TRUE, scale = TRUE)
swidepp403_bl$zcheek = scale(swidepp403_bl$cheek, center = TRUE, scale = TRUE)

# bin next
swidepp403_bin$zbrow = scale(swidepp403_bin$brow, center = TRUE, scale = TRUE)
swidepp403_bin$zcheek = scale(swidepp403_bin$cheek, center = TRUE, scale = TRUE)

# Check that the mean of the new zbrow and zcheek columns is 0 - this is what we want... 
summary(swidepp403_bl)

summary(swidepp403_bin)



```

##  all the data

make data wide, separate bl and bins into different df
```{r}

sdata_score <- data %>%
  pivot_wider(names_from = muscle, 
              values_from = rms) 

# separate bl and bins

sbl_data_score <- sdata_score %>%
  filter(bin == "bin_0")

sbin_data_score <- sdata_score %>%
  filter(bin != "bin_0")

```

Make new col with z score for bl and bin separately
```{r}
# bl first 
sbl_data_zscore <- sbl_data_score %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))
  
summary(sbl_data_zscore)
```


```{r}
# bin next 
sbin_data_zscore <- sbin_data_score %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))
  
summary(sbin_data_zscore)
```

Join bl and bin back together, arrange by trial and bin and drop old brow and cheek

```{r}
sdata_zscore <- bind_rows(sbl_data_zscore, sbin_data_zscore) %>%
  arrange(pp_no, emotion, trial, bin) %>%
  select(-brow, -cheek)
```

##  Difference Scores 

Need to calculate difference scores from baseline. Need to make BIN wide to allow for calculations across columns. Its a bit difficult to do that for both muscles at the same time, so lets separate and work out the difference scores for brow and cheek separately.

### BROW FIRST

#### Add a muscle column 

Just to make things easier to join back together later
```{r}

sbrow_z <- sdata_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, condition, emotion, bin, trial, muscle, Zbrow) 
```

#### Make the bin column wide 

... and rename bin_0 as BL (i.e. baseline)
```{r}
sbrow_z_wide <- sbrow_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zbrow) %>%
  rename(BL = bin_0)

```

#### mutate() diff scores 

Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

```{r}
sbrow_z_diff <- sbrow_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL) %>%
  select(-BL, - starts_with("bin"))
```

#### Make data long again 

```{r}
sbrow_z_diff_long <- sbrow_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin6)
  
```

#### checking on missingness
```{r}
library(naniar)

vis_miss(sbrow_z_diff_long)
```

### NOW CHEEK

#### add a muscle column 

...to make things easier to join back together later
```{r}

scheek_z <- sdata_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, condition, emotion, bin, trial, muscle, Zcheek) 

```

#### Make the bin column wide 

...and rename bin_0 as BL (i.e. baseline)
```{r}

scheek_z_wide <- scheek_z %>%
  pivot_wider(names_from = bin, 
              values_from = Zcheek) %>%
  rename(BL = bin_0)
```

#### mutate() diff scores 

- Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

```{r}
scheek_z_diff <- scheek_z_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL) %>%
  select(-BL, - starts_with("bin"))
```
#### make data long again 

```{r}

scheek_z_diff_long <- scheek_z_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin6)
  
```


# BIND CHEEK AND BROW TOGETHER

```{r}
szdiff_cheek_brow <- bind_rows(scheek_z_diff_long, sbrow_z_diff_long) %>%
  arrange(pp_no, emotion, trial, muscle, bin)

# write to csv 

szdiff_cheek_brow %>%
  write_csv(here("data", "combined", "4.5_zdiff_clean.csv"))

beepr::beep(2)
```


# PLOT distribution of zdiff

```{r}

szdiff_cheek_brow %>%
  group_by(muscle, emotion) %>%
  summarise(meanZdiff = mean(Zdiff, na.rm= TRUE), 
            sdZdiff = sd(Zdiff, na.rm=TRUE))


sep_plot <- szdiff_cheek_brow %>%
  ggplot(aes(x = bin, y = Zdiff)) +
  geom_jitter() +
  facet_grid(vars(muscle), vars(emotion)) +
  labs(title = "Zdiff bl and bin separately")
```
# **COMPARE**

```{r}
library(patchwork)

tog_sep <- tog_plot + sep_plot

print(tog_sep)

```


```{r}

ggsave(here("output", "tog_sep.png"))
```

```{r}
beepr::beep(3)
```
